Fix: LexicalAnalyzer: error when read number literals containing "+" or "-"
Exponent parts of number literals may contain a sign "+" or "-",
for example: 1.23E+12, 1.23E-12, etc.

Bofore this commit,
a sign "+"/"-" of an exponent part of number literals
have been recognized as an addition/subtraction operator,
and for that reason, the lexical analyzer crashed
when it read such number literals.

The above bug has been fixed by this commit.