Fix: the order of arguments had been reversed
The arguments passed to a function in an expression
had been stored in the "args[]" array in reversed order
and then passed to the implementation of the function.

For example, the arguments of the following function call:

    fun(1.25, 2.5, 5.0)

had been stored in the args[] array as follows:

    args[0] = 5.0
    args[1] = 2.5
    args[2] = 1.25

and then they had been passed to the implementation of
the "fun" function, which is connected to Exevalator by
the "connectFunction(...)" method.
This had been very strange and confusing behavior,
and has been fixed by this commit.

After this commit, the arguments of the above
function call are stored in the args[] array as follows:

    args[0] = 1.25
    args[1] = 2.5
    args[2] = 5.0

This change breaks compatibility of function calls,
so the version code of the next release version
will be upgraded to 2.0.