Refactor: Reverse the Order of Return Values in pop_partial_expr_nodes Method
In commit 59ce97410526eea47ae6554b981eef3ecd6aac64,
we added code to reverse the order of elements
returned by the pop_partial_expr_nodes(...) method
on the caller side.

However, in this commit,
we have moved the reversing logic
inside the pop_partial_expr_nodes(...) method itself
and removed the reversing code from the caller side.

This refactoring improves readability and
maintains consistency with the Parser's algorithm
of the Vnano scripting engine,
which is the basis for Exevalator's algorithm.