Fix: workflow: had not considered implicit namespace
In a VB.NET project, at the top of the namespace,
the project name is appended implicitly, by the default settings.
So in "Test" project, the following line in the Test module:

  Imports Rinearn.ExevalatorVB

must be fixed to the following:

  Imports Test.Rinearn.ExevalatorVB

When we compile the test code by vbc command, the above is not necessary.
Hence, we do the above modification dynamically
by "sed" command on the testing workflow.