Remove: TypeScript: connectFunction(...): 'overwrite' arg
For simplicity.

We should be careful for expansion of API.

If we support this 'overwrite' argument,
we must add the same argument to all the other implementations
(of other languages).

We want to keep API as simple as possible,
so decided to suspend this expansion.

If it is truly required from user-side, in future,
we will consider about this expansion again.